refactor everything possible into AdoScalarBase.
